[WEB] update display price

+++ client/web/sudoo/lib/data/api/discovery/request/upsert_product_request.dart
@@ -21,10 +21,10 @@ class UpsertProductRequest extends BaseRequest {
   final DateTime? startDateDiscount;
   final DateTime? endDateDiscount;
   final bool? saleable;
-  final int weight;
-  final int height;
-  final int width;
-  final int length;
+  final double weight;
+  final double height;
+  final double width;
+  final double length;
   final List<UpsertFileRequest>? images;
   final List<String>? categoryIds;
   final UpsertExtrasRequest? extras;
