+++ client/web/sudoo/lib/data/dto/discovery/product_dto.dart
@@ -23,10 +23,10 @@ class ProductDto {
   final DateTime? startDateDiscount;
   final DateTime? endDateDiscount;
   final bool saleable;
-  final int weight;
-  final int height;
-  final int width;
-  final int length;
+  final double weight;
+  final double height;
+  final double width;
+  final double length;
   final List<FileDto> images;
   final SupplierDto supplier;
   final List<CategoryDto> categories;
